+++ CHANGELOG.md
@@ -1605,3 +1605,9 @@ All notable changes to this project will be documented in this file.
 
 - fix error response in message-hook for BepInEx 5.x.x.x version
 - minor optimizations in message-hook
+
+## [2.6.24] - 14.05.2025
+
+- support clear message in handler for in-game runtime
+- properly send keyCode for keys such as arrow keys, space, enter etc for in-game runtime
+- support env files